requirements.txt
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.7.22
charset-normalizer==3.3.0
et-xmlfile==1.1.0
fpdf==1.7.2
idna==3.4
openpyxl==3.1.2
Pillow==10.1.0
python-slugify==8.0.1
PyYAML==6.0.1
requests==2.31.0
slugify==0.0.1
soupsieve==2.5
text-unidecode==1.3
urllib3==2.0.7
